LayoutTests/platform/ios-simulator/editing/pasteboard/paste-noscript-expected.txt

CONSOLE MESSAGE: line 42: TypeError: undefined is not an object (evaluating 'document.getElementById("pastehere").childNodes[5].outerHTML')
This test copies all the elements containing event handlers and javascript urls, pastes them in an editable area and verifies that no script, handlers or javascript urls are copied.
Hello 
CNN Hello  
This is a form
Submit.
Hello 
CNN Hello  
This is a form
Submit.
<button id="button1" onclick="sayHello()" ondblclick="sayHello()" style="width: 100px;">Hello</button>
<div style="text-align: start;"><font face="Times" size="3"><span style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"><button id="button1" style="width: 100px;">Hello</button>&nbsp;</span></font></div>
<a id="anchor1" href="http://www.cnn.com/">CNN</a>
undefined
<a id="anchor2" href="javascript:sayHello()">Hello</a>
